moved background merging code to a separate utility class
